demo/requirements.txt
pillow==3.1.1
django==1.9.4
django-tinymce==2.3.0
django-taggit==0.18.0
django-taggit-templatetags==0.4.6dev
coverage==4.0.3
django-coverage==1.2.4
factory_boy==2.6.1

# packages required by ostinato, should be same versions as in ostinato
# setup.py file

django-mptt==0.8.3

# django-appregister==0.3.1
git+http://github.com/andrewebdev/django-appregister.git#egg=django-appregister

